+++ packages/common/src/task/interval.ts
@@ -1,0 +1,7 @@
+export const getSetInterval = (
+  callback: () => void,
+  delay?: number | undefined,
+): (() => void) => {
+  const timeout = setInterval(callback, delay);
+  return () => clearInterval(timeout);
+};